CMake toolchain file and listfile modifications for mingw32
The toolchain file can be used together with the changes to the CMake
listfile in this commit to cross-compile ia for windows with mingw32. A
command like the following can be run from the build directory to
compile a 32-bit windows ia.exe from Linux:

cmake -DWIN32=TRUE -DMSVC=FALSE -DARCH=32bit \
      -DCMAKE_TOOLCHAIN_FILE=../Toolchain-cross-mingw32.txt ..

The listfile changes clean up the setting and checking of the ARCH
variable a bit to allow compilation for either 64bit or 32bit so that
the command-line value will override the automatic detection (thus
allowing 64-bit machines to compile for 32-bit).

The other listfile change is to statically link libgcc and libstdc++6
when using GCC, so that these don't have to be included on their own as
dlls in the release package. For Clang, statically linking libstdc++6
might also be a good idea, but I haven't tested this.

The toolchain file checks the ARCH variable to set the correct mingw32
compiler. I'm not sure if this is the best CMake setup for
cross-compilation, but it works when compiling for both architectures.