Fixed weird behavior for burning map features
Since the map feature new turn hooks run top to bottom, left to right,
spreading fire behaved differently depending on direction. E.g. if
fire spread right, it would burn any actor in that cell immediately,
but if it spread left, it would not burn the actor there until the
next turn.

Now burning map features can never spread fire or damage actors etc
on the same turn that the map feature started burning. This is more
fair to the player, and the behavior is the same regardless of which
direction the fire is spreading.